second

+++ resources/views/posts/show.blade.php
@@ -26,16 +26,18 @@
                     <dd>{{ date('j M Y H:i', strtotime($post->updated_at)) }}</dd>
                 </dl>
                 <hr>
-                <div class="row">
-                    <div class="col-sm-6">
-                        {!! Html::linkRoute('posts.edit', 'Edit', array($post->id), array('class' => 'btn btn-warning btn-block')) !!}
+                @if($post->user_id == Auth::user()->id)
+                    <div class="row">
+                        <div class="col-sm-6">
+                            {!! Html::linkRoute('posts.edit', 'Edit', array($post->id), array('class' => 'btn btn-warning btn-block')) !!}
+                        </div>
+                        <div class="col-sm-6">
+                            {!! Form::open(['route' => ['posts.destroy', $post->id], "method" => "DELETE"]) !!}
+                            {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-block']) !!}
+                            {!! Form::close() !!}
+                        </div>
                     </div>
-                    <div class="col-sm-6">
-                        {!! Form::open(['route' => ['posts.destroy', $post->id], "method" => "DELETE"]) !!}
-                        {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-block']) !!}
-                        {!! Form::close() !!}
-                    </div>
-                </div>
+                @endif
                 <div class="row">
                     <div class="col-md-12">
                         {{ Html::linkRoute('posts.index', '<< See All Posts', [], ['class' => 'btn btn-success btn-block', 'style' => 'margin: 15px 0 0 0;']) }}
